[FIX] sale_blanket_order: qty change keep bol
Actually it makes more sense to not change blanket order when quantity
changes as you can end up in situation when blanket order is simply
removed, when all quantity is used or more is used. So in these cases
it would remove all data showing how much of that blanket order was
consumed, which is wrong.